requirements.txt
#Required modules
streamlit
google-generativeai
python-dotenv
